Create stringCompressionI.js
You are given a string S and you have to implement a function to perform basic string compression, using the counts of repeated characters. Return a string containing single appearance of every character which is being repeated along with the number it is being repeated.

If the "compressed" string would not become smaller than the original string, your function should return the original string.

You can assume the string has only uppercase and lowercase letters (a - z).