Create matrixTraversal.js
A matrix and a valid current position (row number and column number) are provided as input. Given a direction (RIGHT is 1, DOWN in 2, LEFT is 3 and UP is 4) and number of steps to move, output the values along the path taken to the new position, including the value at the new position.

If any position during traversal is out of bounds, output a single -1 and not the values along the path.